text animation moved into index.tsx of Carousel
